# docs/openapi/replicadb-server.json
{
  "components": {
    "responses": {
      "BadRequestProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The request is malformed or fails validation."
      },
...
      "ForbiddenProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The authenticated identity lacks the required permission."
      },
      "InternalServerErrorProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The server could not complete the request."
      },
...
      "TooManyRequestsProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The request is throttled after repeated authentication failures."
      },
      "UnauthorizedProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "Authentication is required or credentials are invalid."
      }
    },
    "schemas": {
...
      "CsrfTokenResponse": {
        "description": "CSRF bootstrap values issued by Spring Security.",
        "properties": {
          "headerName": {
            "description": "HTTP header name expected on protected mutations.",
            "example": "X-XSRF-TOKEN",
            "type": "string"
          },
          "parameterName": {
            "description": "Framework request parameter name.",
            "type": "string"
          },
          "token": {
            "description": "Opaque CSRF value that must match the XSRF-TOKEN cookie.",
            "readOnly": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "KeyringEnvelopeCountResponse": {
        "properties": {
          "count": {
            "format": "int64",
            "type": "integer"
          },
          "keyVersion": {
            "type": "string"
          },
          "known": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "KeyringReencryptResponse": {
        "properties": {
          "converged": {
            "type": "boolean"
          },
          "reencrypted": {
            "format": "int32",
            "type": "integer"
          },
          "remaining": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "KeyringStatusResponse": {
        "properties": {
          "converged": {
            "type": "boolean"
          },
          "currentVersion": {
            "type": "string"
          },
          "envelopes": {
            "items": {
              "$ref": "#/components/schemas/KeyringEnvelopeCountResponse"
            },
            "type": "array"
          },
          "knownVersions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "reencryptionRequired": {
            "format": "int64",
            "type": "integer"
          },
          "unknownVersionCount": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "LoginRequest": {
        "description": "Credentials used only to create a server-owned session.",
        "properties": {
          "password": {
            "description": "Password accepted only for authentication and never returned.",
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "description": "ReplicaDB username.",
            "example": "api-operator",
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807 problem response. Dynamic detail text is credential-redacted.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "format": "uri-reference",
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UserIdentityResponse": {
        "description": "Identity associated with the current authenticated session.",
        "properties": {
          "id": {
            "description": "User identifier.",
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "description": "Global authorization role.",
            "enum": [
              "ADMIN",
              "OPERATOR",
              "VIEWER"
            ],
            "type": "string"
          },
          "username": {
            "description": "Current username.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF header for protected state-changing requests. Its value matches the XSRF-TOKEN cookie.",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "Session cookie issued by an authenticated, same-origin ReplicaDB Server deployment.",
        "in": "cookie",
        "name": "JSESSIONID",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Static contract for the authenticated ReplicaDB Server control plane. Use it against a same-origin deployment that owns the session and CSRF cookies.",
    "title": "ReplicaDB Server API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/auth/csrf": {
      "get": {
        "description": "Initializes the XSRF-TOKEN cookie and returns the framework header and parameter names. Call this public endpoint before protected state-changing requests.",
        "operationId": "getCsrfToken",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfTokenResponse"
                }
              }
            },
            "description": "CSRF contract initialized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorProblem"
          }
        },
        "summary": "Initialize CSRF protection",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticates a user and establishes the server-owned session. Login is public and exempt from CSRF, but database-backed throttling limits repeated failures by account and source address.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdentityResponse"
                }
              }
            },
            "description": "Authenticated identity returned and session established"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestProblem"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequestsProblem"
          }
        },
        "summary": "Create an authenticated session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Invalidates the current server session and clears its security context. The session cookie and CSRF header are required.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Session ended"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "End the authenticated session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Returns the user identifier, username, and global role associated with the current session.",
        "operationId": "getCurrentIdentity",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdentityResponse"
                }
              }
            },
            "description": "Current identity returned"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get the current identity",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/keyring/reencrypt": {
      "post": {
        "description": "Re-encrypts one bounded batch using the current key version. Repeat until remaining is zero. ADMIN and CSRF are required.",
        "operationId": "reencryptKeyring",
        "parameters": [
          {
            "description": "Maximum number of envelopes to process in this request.",
            "in": "query",
            "name": "batchSize",
            "required": false,
            "schema": {
              "default": "200",
              "description": "Maximum number of envelopes to process in this request.",
              "maximum": 1000,
              "minimum": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/KeyringReencryptResponse"
                }
              }
            },
            "description": "Re-encryption batch completed"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Re-encrypt datasource envelopes",
        "tags": [
          "Keyring"
        ]
      }
    },
    "/api/v1/keyring/status": {
      "get": {
        "description": "Returns known key versions and encrypted datasource envelope counts without key material. ADMIN is required.",
        "operationId": "getKeyringStatus",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/KeyringStatusResponse"
                }
              }
            },
            "description": "Keyring status returned"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get keyring status",
        "tags": [
          "Keyring"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}